Remove dead code related to TAS
- Removes old SSO implementation
- Removes old login and registration pages
- Removes GENI federation code
- Removes TAS admin UI and password reset / forgot username flows

This also fixes a bug where migration of federated accounts and specifically
cinder volumes on KVM would fail because there was no Keystone token on-hand
for the user. We need to ensure the user provides their old TAS password so
we can log them in to their old account to do this. Now the migration page
will detect this case and prompt the user for the legacy password and handle
regenerating these tokens for them.
