+++ chameleon/templates/federation/confirm_legacy_credentials.html
@@ -2,18 +2,20 @@
 {% load bootstrap3 %}
 {% block title %}Credentials Verification{% endblock %}
 {% block content %}
-<h1>Credential Verification</h1>
+<h1>Confirm legacy credentials</h1>
 
 <div class="row">
   <div class="col-sm-6">
     <p>
-     Please verify your credentials before proceeding.
+      In order to perform some of the data migrations from your legacy account,
+      the system needs your legacy account credentials.
+    </p>
+    <p>
+      Please enter your old Chameleon username and password to continue.
     </p>
     <form method="post">
       {% csrf_token %}
-
       {% bootstrap_form form %}
-
       <div class="actions">
         <button type="submit" class="btn btn-primary">Submit</button>
       </div>
