--- chameleon/templates/registration/login.html
@@ -1,55 +1,0 @@
-{% extends "layout/default.html" %}
-{% load bootstrap3 chameleon_openid_tags %}
-{% block title %}Log In{% endblock %}
-{% block content %}
-  {% if request.user.is_authenticated %}
-  <h2>You're already logged in!</h2>
-
-  <p class="lead">
-    This is odd, you are already logged in as <b>{{request.user}}</b>!
-    <a href="{% url 'dashboard' %}">Click here to head back to the User Dashboard.</a>
-  </p>
-  {% else %}
-  <h2>Login to Chameleon</h2>
-
-  <div class="row">
-    <div class="col-sm-6">
-      <form method="post">
-        {% csrf_token %}
-
-        {% bootstrap_form form %}
-
-        <div class="form-group">
-          <button type="submit" class="btn btn-lg btn-success">Log in</button>
-          <a class="btn btn-link" href="{% url 'tas:password_reset' %}">Forgot your password?</a>
-          <a class="btn btn-link" href="{% url 'tas:recover_username' %}">Forgot your username?</a>
-        </div>
-      </form>
-
-      <h3>Other login options</h3>
-      {% openid_login_options %}
-    </div>
-
-
-    <div class="col-sm-6">
-      <div class="well">
-        <h3>Register for an Account</h3>
-        <p class="lead">
-          Don't have a Chameleon account? <a href="{% url 'register' %}" class="btn btn-default">Register for an Account</a>
-        </p>
-        <p class="lead">
-          If you have a TACC account you can use your TACC credentials to log in to
-          the Chameleon portal and activate your Chameleon account.
-        </p>
-      </div>
-    </div>
-  </div>
-
-  {% endif %}
-{% endblock %}
-
-{% block scripts %}
-<script>
-jQuery(function($) { $('#id_username').focus(); });
-</script>
-{% endblock %}
