--- chameleon/templates/sso/keystone_login_unavailable.html
@@ -1,17 +1,0 @@
-{% extends "layout/default.html" %}
-{% load bootstrap3 %}
-{% block title %}Account Error{% endblock %}
-{% block content %}
-<h1>No Chameleon Project Found</h1>
-
-<div class="row">
-  <div class="col-sm-6">
-    <p>
-     A Chameleon project associated with your account was not found. If your account was recently updated with a project affiliation, please wait a few minutes for the account to update before trying again.
-    </p>
-    <p>
-     To create or join a project, please follow the steps <a href="https://chameleoncloud.readthedocs.io/en/latest/getting-started/index.html#step-2-create-or-join-a-project" target="_blank">here</a>.
-    </p>
-  </div>
-</div>
-{% endblock %}
