--- chameleon/templates/sso/sso_callback_template.html
@@ -1,22 +1,0 @@
-<!DOCTYPE html>
-<html xmlns="http://www.w3.org/1999/xhtml">
-  <head>
-    <title>Keystone WebSSO redirect</title>
-  </head>
-  <body>
-     <form id="sso" name="sso" action="{{ sso_url }}" method="post">
-       Please wait...
-       <br/>
-       <input type="hidden" name="token" id="token" value="{{ sso_token }}"/>
-       <noscript>
-         <input type="submit" name="submit_no_javascript" id="submit_no_javascript"
-            value="If your JavaScript is disabled, please click to continue"/>
-       </noscript>
-     </form>
-     <script type="text/javascript">
-       window.onload = function() {
-         document.forms['sso'].submit();
-       }
-     </script>
-  </body>
-</html>
