--- chameleon_openid/templates/chameleon_openid/activate_geni.html
@@ -1,46 +1,0 @@
-{% extends "layout/default.html" %}
-
-{% block title %}Activate Chameleon-GENI Federation Access{% endblock %}
-
-{% block content %}
-<h2>Activate Chameleon-GENI Federation Access</h2>
-
-<div class="row">
-  <div class="col-md-9">
-    <p>
-      As a member of the <a href="{{geni.url}}">{{geni.name}}</a> project on the GENI Portal,
-      you have access to use Chameleon. Click <strong>Activate Access</strong> below to be
-      automatically added to the {{chameleon.name}} project on Chameleon, through which you
-      will have access to use Chameleon resources.
-    </p>
-    <p>
-      The {{chameleon.name}} project will give you access to test the capabilities that
-      Chameleon offers. Because the resources available via this project are shared, once
-      you are ready to scale up your project to fully utilize Chameleon we ask that you
-      <a href="{% url 'projects:create_project' %}">submit a project request</a>.
-    </p>
-  </div>
-</div>
-
-<hr>
-
-<form method="post">
-  {% csrf_token %}
-
-  <div class="checkbox required">
-    <label>
-      <input type="checkbox" name="accept_user_terms" required>
-      I agree to abide by Chameleon Acceptable Use Policies
-    </label>
-    <p class="help-block">
-      Please review the Chameleon <a href="{% url 'tc_view_specific_page' 'user-terms' %}">User Terms of Use</a>
-      before proceeding.
-    </p>
-  </div>
-
-  <button type="submit" class="btn btn-primary btn-lg">Activate Access</button>
-  &nbsp;
-  <a class="btn btn-link" href="{% url 'dashboard' %}">Cancel</a>
-</form>
-
-{% endblock %}
