--- chameleon_openid/templates/chameleon_openid/connect.html
@@ -1,38 +1,0 @@
-{% extends "layout/default.html" %}
-{% load bootstrap3 %}
-{% block title %}OpenID Connect{% endblock %}
-{% block content %}
-<h1>OpenID Connect</h1>
-
-<p class="lead">
-  Hello, {{ openid.ax.full_name.0 }}! You have successfully authenticated with the remote
-  identity:<br><code>{{ openid.url }}</code>
-</p>
-
-<p class="lead">
-  It looks like you already have a Chameleon account, but this is the first time you've
-  logged in using this OpenID Identity. Please enter your Chameleon credentials below to
-  connect this identity with your existing Chameleon account.
-</p>
-
-<!-- OpenID Debug Info {{openid}} -->
-
-<div class="row">
-  <div class="col-sm-9">
-    <h2>Connect account</h2>
-    <p>
-      Enter your Chameleon username and password to link your Chameleon account with this
-      OpenID identity.
-    </p>
-    <form method="post">
-      {% csrf_token %}
-      {% bootstrap_form form %}
-      <div class="actions">
-        <button type="submit" class="btn btn-primary">Connect account</button>
-      </div>
-    </form>
-  </div>
-</div>
-
-
-{% endblock %}
