--- chameleon_openid/templates/chameleon_openid/connect_or_register.html
@@ -1,50 +1,0 @@
-{% extends "layout/default.html" %}
-{% load bootstrap3 %}
-{% block title %}OpenID: Connect Account{% endblock %}
-{% block content %}
-<h1>OpenID: Connect your Account</h1>
-
-<p class="lead">
-  Hello, {{ openid.ax.full_name.0 }}! You have successfully authenticated with the remote
-  identity:<br><code>{{ openid.url }}</code>
-</p>
-
-<p class="lead">
-  However, we were unable to match this to a Chameleon user account! Please either connect
-  this identity with your existing Chameleon account, or create a Chameleon account if you
-  don't already have one!
-</p>
-
-<!-- OpenID Debug Info {{openid}} -->
-
-<div class="row">
-  <div class="col-sm-6">
-    <h2>Connect account</h2>
-    <p>
-      Enter your <strong>Chameleon username and password</strong> to link your Chameleon
-      account with this OpenID identity. If you have a <strong>TACC account</strong>, you
-      can use your TACC username and password to activate your Chameleon account and link
-      your OpenID at the same time.
-    </p>
-    <form method="post">
-      {% csrf_token %}
-
-      {% bootstrap_form form %}
-
-      <div class="actions">
-        <button type="submit" class="btn btn-primary">Connect account</button>
-      </div>
-    </form>
-  </div>
-  <div class="col-sm-6">
-    <h2>Register account</h2>
-    <p>
-      If you don't have a Chameleon account, please register for one. After registration
-      is complete, we will link your OpenID identity to your account automatically.
-    </p>
-    <a class="btn btn-info" href="{% url 'chameleon_openid:openid_register' %}">Register new account</a>
-  </div>
-</div>
-
-
-{% endblock %}
