--- chameleon_openid/templates/chameleon_openid/login.html
@@ -1,7 +1,0 @@
-{% extends "layout/default.html" %}
-{% load bootstrap3 chameleon_openid_tags %}
-{% block title %}OpenID Log In{% endblock %}
-{% block content %}
-<h1>OpenID Login</h1>
-{% openid_login_options %}
-{% endblock %}
