--- chameleon_openid/test_fixtures/registration_form.json
@@ -1,15 +1,0 @@
-{
-  "firstName": "John",
-  "lastName": "Doe",
-  "email": "jdoe1@tacc.utexas.edu",
-  "phone": "555 555 1234",
-  "institutionId": 1,
-  "institution": "",
-  "departmentId": 127,
-  "title": "University Research Staff",
-  "countryId": 230,
-  "citizenshipId": 230,
-  "username": "jdoe1",
-  "password": "asdf;1234;",
-  "confirmPassword": "asdf;1234;"
-}
