--- tas/templates/admin/tas/reset_user_password.html
@@ -1,52 +1,0 @@
-{% extends "admin/base_site.html" %}
-{% load i18n admin_static %}
-{% load admin_urls %}
-
-{% block extrahead %}{{ block.super }}
-<script type="text/javascript" src="{% url 'admin:jsi18n' %}"></script>
-{% endblock %}
-{% block extrastyle %}{{ block.super }}<link rel="stylesheet" type="text/css" href="{% static "admin/css/forms.css" %}" />{% endblock %}
-{% block bodyclass %}{{ block.super }} {{ opts.app_label }}-{{ opts.model_name }} change-form{% endblock %}
-{% if not is_popup %}
-{% block breadcrumbs %}
-<div class="breadcrumbs">
-<a href="{% url 'admin:index' %}">{% trans 'Home' %}</a>
-&rsaquo; <a href="{% url 'admin:app_list' app_label=opts.app_label %}">{{ opts.app_config.verbose_name }}</a>
-&rsaquo; <a href="{% url opts|admin_urlname:'changelist' %}">{{ opts.verbose_name_plural|capfirst }}</a>
-&rsaquo; {{title}}
-</div>
-{% endblock %}
-{% endif %}
-{% block content %}<div id="content-main">
-<form action="{{ form_url }}" method="post" id="{{ opts.model_name }}_form">{% csrf_token %}{% block form_top %}{% endblock %}
-<div>
-{% if is_popup %}<input type="hidden" name="_popup" value="1" />{% endif %}
-
-<p>{% blocktrans %}Are you sure you want to send <strong>Password Reset Notifications</strong> to the following users?{% endblocktrans %}</p>
-
-<fieldset class="module aligned">
-  <div class="form-row">
-    <label>Selected Users:</label>
-    <ul>
-    {% for user in selected_users %}
-      <li>
-        <input type="hidden" name="{{ action_checkbox_name }}" value="{{ user.id }}" />
-        {{user.first_name}} {{user.last_name}} ({{user.username}})
-      </li>
-    {% endfor %}
-    </ul>
-  </div>
-
-</fieldset>
-
-<input type="hidden" name="action" value="reset_user_password" />
-<input type="hidden" name="post" value="yes" />
-
-<div class="submit-row">
-  <input type="submit" value="{% trans 'Send Password Reset Notification(s)' %}" class="default" />
-  <input type="button" onclick="window.history.back(); return false;" value="{% trans 'No, take me back' %}" />
-</div>
-
-</div>
-</form></div>
-{% endblock %}
