--- tas/templates/tas/email_confirmation.html
@@ -1,24 +1,0 @@
-{% extends "layout/default.html" %}
-
-{% load bootstrap3 %}
-
-{% block title %}Email Confirmation{% endblock %}
-
-{% block content %}
-<h2>Email Confirmation</h2>
-
-<p class="lead">
-  Almost finished! Enter the verification code that you received via email
-  and the username you registered with to activate your account.
-</p>
-
-<form method="post">
-  {% csrf_token %}
-
-  {% bootstrap_form form %}
-
-  <div class="form-group">
-    <button type="submit" class="btn btn-success">Verify</button>
-  </div>
-</form>
-{% endblock %}
