+++ tas/templates/tas/email_subscription_opt_in.html
@@ -1,24 +1,36 @@
 <html>
   <head>
-      <style>
-          body {font-family: Cambria, Cochin, Georgia, Times, 'Times New Roman', serif}
-      </style>
+    <style>
+        body {
+          font-family: Cambria, Cochin, serif;
+        }
+    </style>
   </head>
   <body>
-    <h3>Dear {{fname}},</h3>
+    <h3>Dear {{ fname }},</h3>
 
     <p>Welcome to Chameleon – we’re excited to have you join us!</p>
 
     <p>
-    We’d like to draw your attention to the user icon in the top right corner of <a href="https://www.chameleoncloud.org" target="_blank">www.chameleoncloud.org</a>: once you’re added to a Chameleon project, it provides a shortcut to the Chameleon GUI via the <a href="https://www.chameleoncloud.org/docs/getting-started/experiment-quickstart/" target="_blank">experiment option</a>.
-
-    If at any point you need help please submit a ticket via the <a href="https://www.chameleoncloud.org/user/help/ticket/new" target="_blank">help desk</a>.
-
-    And now the fun begins: click on our <a href="https://chameleoncloud.readthedocs.io/en/latest/getting-started/index.html" target="_blank">getting started guide</a> and give it a whirl!
+      We’d like to draw your attention to the user icon in the top right corner
+      of <a href="https://www.chameleoncloud.org"
+      target="_blank">www.chameleoncloud.org</a>: once you’re added to a
+      Chameleon project, it provides a shortcut to the Chameleon GUI via the <a
+      href="https://www.chameleoncloud.org/docs/getting-started/experiment-quickstart/"
+      target="_blank">experiment option</a>.
     </p>
-
     <p>
-    Happy experimenting, we are looking forward to working with you!
+      If at any point you need help please submit a ticket via the <a
+      href="https://www.chameleoncloud.org/user/help/ticket/new"
+      target="_blank">help desk</a>.
+    </p>
+    <p>
+      And now the fun begins: click on our <a
+      href="https://chameleoncloud.readthedocs.io/en/latest/getting-started/index.html"
+      target="_blank">getting started guide</a> and give it a whirl!
+    </p>
+    <p>
+      Happy experimenting, we are looking forward to working with you!
     </p>
   </body>
 </html>
