--- tas/templates/tas/password_reset.html
@@ -1,21 +1,0 @@
-{% extends "layout/default.html" %}
-
-{% load bootstrap3 %}
-
-{% block title %}Password Reset{% endblock %}
-
-{% block content %}
-<h2>Password Reset</h2>
-
-<p class="lead">{{ message }}</p>
-
-<form method="post">
-  {% csrf_token %}
-
-  {% bootstrap_form form %}
-
-  {% buttons %}
-    <button type="submit" class="btn btn-success">Request Password Reset</button>
-  {% endbuttons %}
-</form>
-{% endblock %}
