--- tas/templates/tas/recover_username.html
@@ -1,15 +1,0 @@
-{% extends "layout/default.html" %}
-{% load staticfiles chameleon_tags bootstrap3 %}
-{% block title %}Forgot Username{% endblock %}
-{% block content %}
-
-    <form method="post">
-      {% csrf_token %}
-              {% bootstrap_form form %}
-
-              <div class="form-group">
-        <button type="submit" class="btn btn-success">Submit</button>
-        <a href="{% url 'tas:profile' %}" class="btn btn-default">Cancel</a>
-      </div>
-    </form>
-{% endblock %}
